LayoutTests/inspector/animation/effectChanged-expected.txt

Tests for the Animation.effectChanged event.


== Running test suite: Animation.effectChanged
-- Running test case: Animation.effectChanged.NewEffect
PASS: Animation should have an effect.
Changing effect...

PASS: Animation should have an effect.
PASS: Animation effect should have changed.

-- Running test case: Animation.effectChanged.NullEffect
PASS: Animation should have an effect.
Changing effect...

PASS: Animation should not have an effect.
PASS: Animation effect should have changed.
